refactor: improve some codes
The "getPost" function in the "postController.js" file, located in the "/controllers/user/" directory, previously used the ".findOne" function to retrieve post information. However, to increase the view count of the related post, the ".findOne" function has been replaced with the ".findOneAndUpdate" function.
Additionally, during the testing stage, the "required" property of the "author" field in the post model schema has been temporarily changed from "true" to "false".